ai-agents/ai-travel-agent/requirements.txt
streamlit 
phidata
openai
google-search-results
lxml_html_clean
googlesearch-python
pycountry
google-generativeai
ollama